+++ lsteamclient/cppISteamNetworkingUtils_SteamNetworkingUtils002.h
@@ -17,7 +17,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_GetLocalPingLocatio
 struct cppISteamNetworkingUtils_SteamNetworkingUtils002_EstimatePingTimeBetweenTwoLocations_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     const SteamNetworkPingLocation_t *location1;
     const SteamNetworkPingLocation_t *location2;
 };
@@ -26,7 +26,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_EstimatePingTimeBet
 struct cppISteamNetworkingUtils_SteamNetworkingUtils002_EstimatePingTimeFromLocalHost_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     const SteamNetworkPingLocation_t *remoteLocation;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_EstimatePingTimeFromLocalHost( struct cppISteamNetworkingUtils_SteamNetworkingUtils002_EstimatePingTimeFromLocalHost_params *params );
@@ -36,7 +36,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils002_ConvertPingLocationToStr
     void *linux_side;
     const SteamNetworkPingLocation_t *location;
     char *pszBuf;
-    int cchBufSize;
+    int32_t cchBufSize;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_ConvertPingLocationToString( struct cppISteamNetworkingUtils_SteamNetworkingUtils002_ConvertPingLocationToString_params *params );
 
@@ -60,7 +60,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_CheckPingDataUpToDa
 struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPingToDataCenter_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID popID;
     SteamNetworkingPOPID *pViaRelayPoP;
 };
@@ -69,7 +69,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPingToDataCenter
 struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetDirectPingToPOP_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID popID;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_GetDirectPingToPOP( struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetDirectPingToPOP_params *params );
@@ -77,16 +77,16 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_GetDirectPingToPOP(
 struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPOPCount_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPOPCount( struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPOPCount_params *params );
 
 struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPOPList_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID *list;
-    int nListSz;
+    int32_t nListSz;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPOPList( struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetPOPList_params *params );
 
@@ -111,7 +111,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils002_SetConfigValue_params
     bool _ret;
     uint32_t eValue;
     uint32_t eScopeType;
-    intptr_t scopeObj;
+    int32_t scopeObj;
     uint32_t eDataType;
     const void *pArg;
 };
@@ -123,7 +123,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils002_GetConfigValue_params
     uint32_t _ret;
     uint32_t eValue;
     uint32_t eScopeType;
-    intptr_t scopeObj;
+    int32_t scopeObj;
     uint32_t *pOutDataType;
     void *pResult;
     size_t *cbResult;
